backend/requirements.txt
Flask==3.0.3
Werkzeug==3.1.3
flask-cors==3.0.10
requests==2.26.0
python-dotenv==0.19.1
playwright==1.17.0
pyee==8.1.0
Pillow==10.2.0
pandas==2.2.0
httpx==0.27.0
beautifulsoup4==4.12.2
openpyxl==3.1.2
openai==0.28.1
